+++ api/RestRequest.php
@@ -1,0 +1,106 @@
+<!--
+                           _    _______          _     
+                          | |  |__   __\        | |    
+                     _ __ | |_ __ | | ___   ___ | |___ 
+                    | '_ \| | '_ \| |/ _ \ / _ \| / __|
+                    | | | | | |_) | | (_) | (_) | \__ \
+                    |_| |_|_| .__/|_|\___/ \___/|_|___/
+ ___________________________| |_________________________________________
+|                           |_|                                        |\
+|                                                                      |_\
+|   File    : RestRequest.php                                             |
+|   Created : 07-Mar-2013                                                 |
+|   By      : atrilla                                                     |
+|                                                                         |
+|   nlpTools - Natural Language Processing Toolkit for PHP                |
+|                                                                         |
+|   Copyright (c) 2013 Alexandre Trilla                                   |
+|                                                                         |
+|   ___________________________________________________________________   |
+|                                                                         |
+|   This file is part of nlpTools.                                        |
+|                                                                         |
+|   nlpTools is free software: you can redistribute it and/or modify      |
+|   it under the terms of the MIT/X11 License as published by the         |
+|   Massachusetts Institute of Technology. See the MIT/X11 License        |
+|   for more details.                                                     |
+|                                                                         |
+|   You should have received a copy of the MIT/X11 License along with     |
+|   this source code distribution of nlpTools (see the COPYING file       |
+|   in the root directory). If not, see                                   |
+|   <http://www.opensource.org/licenses/mit-license>.                     |
+|_________________________________________________________________________|
+-->
+
+<?php
+
+/**
+ * @brief Handles REST requests.
+ * See: http://www.gen-x-design.com/archives/create-a-rest-api-with-php
+ *
+ * @author Alexandre Trilla (atrilla)
+ */
+class RestRequest  
+{  
+    private $request_vars;  
+    private $http_accept;  
+    private $method;  
+    private $service;
+    private $text;
+  
+    public function __construct()  
+    {  
+        $this->request_vars      = array();  
+        $this->service           = '';  
+        $this->text              = '';  
+        $this->http_accept       = (strpos($_SERVER['HTTP_ACCEPT'], 'json')) ? 'json' : 'xml';  
+        $this->method            = 'get';  
+    }  
+  
+    public function setService($s)  
+    {  
+        $this->service = $s;  
+    }  
+
+    public function setTextToProc($t)  
+    {  
+        $this->text = $t;  
+    }  
+  
+    public function setMethod($method)  
+    {  
+        $this->method = $method;  
+    }  
+  
+    public function setRequestVars($request_vars)  
+    {  
+        $this->request_vars = $request_vars;  
+    }  
+  
+    public function getService()  
+    {  
+        return $this->service;  
+    }  
+
+    public function getTextToProc()  
+    {  
+        return $this->text;  
+    }  
+  
+    public function getMethod()  
+    {  
+        return $this->method;  
+    }  
+  
+    public function getHttpAccept()  
+    {  
+        return $this->http_accept;  
+    }  
+  
+    public function getRequestVars()  
+    {  
+        return $this->request_vars;  
+    }  
+}
+
+?>